% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffvb.R
\name{ffvb}
\alias{ffvb}
\title{Fixed Form Variational Bayes Algorithm}
\usage{
ffvb(y)
}
\arguments{
\item{y}{a vector}
}
\value{
vector with components \mu_\mu, \tau_\mu, \alpha_\tau, \beta_\tau
}
\description{
Fixed Form Variational Bayes Algorithm
}
\examples{
y<-c(1,2,3,4,5,4,3,5)
ffvb(y)
}
